code to cc
