+++ source/common/formatter/BUILD
@@ -36,6 +36,7 @@ envoy_cc_library(
 envoy_cc_library(
     name = "substitution_format_string_lib",
     hdrs = ["substitution_format_string.h"],
+    srcs = ["substitution_format_string.cc"],
     deps = [
         ":substitution_formatter_lib",
         "//source/common/config:datasource_lib",
